feat: Updated REAMDE, minor changes
