+++ docs/adding_new_rule.md
@@ -1,0 +1,3 @@
+## Adding new rule
+
+TODO
